[Fixbug] customer page

+++ admin/customer/new.php
@@ -9,7 +9,21 @@
                 <div class="content-panel">
                     <br>
                     <?php
-                    echo "<h5 class='notifications'>$MESSAGE</h5>";
+                    if (!empty($MESSAGE) && $MESSAGE == 'Thêm mới thành công!') {
+                        echo '<script>
+                            Swal.fire({
+                                title: "Thêm mới thành công!",
+                                icon: "success",
+                                showConfirmButton: false,
+                                timer: 1500
+                            });
+                            setTimeout(function() {
+                                window.location.href = "index.php?btn_list";
+                            }, 1500);
+                        </script>';
+                    } else {
+                        echo "<h5 class='notifications'>$MESSAGE</h5>";
+                    }
                     ?>
                     <form id="form" action="index.php" method="POST" enctype="multipart/form-data">
                         <div class="form-group">
